requirements.txt
numpy>=1.20.0
pyyaml>=5.4.1
matplotlib>=3.4.2
networkx>=2.5

